+++ src/satisfies-optionDefinitions.js
@@ -5,16 +5,17 @@ const pkg = require('../package.json');
 // Todo: We really need a comamnd-line-args-TO-typedef-jsdoc generator!
 /* eslint-disable jsdoc/require-property */
 /**
-* @typedef {PlainObject} GetLicenseTypeOptions
+* @typedef {PlainObject} SatisfiesOptions
 */
 /* eslint-enable jsdoc/require-property */
 
 const optionDefinitions = [
   {
-    name: 'licenseExpression', type: String, defaultOption: true, alias: 'l',
-    description: 'License expression for which to obtain a type (degree of ' +
-      'permissiveness)',
-    typeLabel: '{underline licenseExpression}'
+    name: 'licenseExpressions', type: String,
+    multiple: true, defaultOption: true, alias: 'l',
+    description: 'License expressions, the first of which is checked ' +
+      'against the second to see if it is satisfied by it.',
+    typeLabel: '{underline licenseExpressions}'
   }
 ];
 
@@ -22,7 +23,7 @@ const cliSections = [
   {
     // Add italics: `{italic textToItalicize}`
     content: pkg.description +
-      '\n\n{italic get-license-type licenseExpressions}'
+      '\n\n{italic satisfies expr1 expr2}'
   },
   {
     optionList: optionDefinitions